src/requirements.txt
azure-core==1.30.2
azure-identity==1.17.1
azure-storage-blob==12.22.0
fastapi==0.112.0
PyMySQL==1.1.1
python-multipart==0.0.9
sqlmodel==0.0.21
uvicorn==0.30.5
